+++ client/src/pages/ClientOrder/index.js
@@ -1,126 +1,107 @@
 import React, { Component } from "react";
 import axios from "axios";
+import {Button} from 'react-materialize';
+import {Input} from 'react-materialize';
+import {Row} from 'react-materialize';
+import "./style.css";
 
-class DriverSignup extends Component {
+class ClientOrder extends Component {
 
     state = {
-          firstName: "",
-          lastName: "",
-          email: "",
-          phoneNumber: "",
-          username: "",
-          password: ""
+          startLocation: "",
+          endLocation: "",
+          loadDescription: "",
+          vehicleRequirement: "",
+          status: "",
+          clientId: "",
+          driverId: ""
       };
     
     handleInputChange = event => {
         // Getting the value and name of the input which triggered the change
         const { name, value } = event.target;
-    
         // Updating the input's state
         this.setState({
             [name]: value
         });
     };
-  
+    
     handleFormSubmit = event => {
         // Preventing the default behavior of the form submit (which is to refresh the page)
         event.preventDefault();
-        const newClient = {
-            firstName: this.state.firstName,
-            lastName: this.state.lastName,
-            email: this.state.email,
-            phoneNumber: this.state.phoneNumber,
-            username: this.state.username,
-            password:this.state.password
+        const thisState = this;
+        const newOrder = {
+            startLocation: this.state.startLocation,
+            endLocation: this.state.endLocation,
+            loadDescritpion: this.state.loadDescritpion,
+            vehicleRequirement: this.state.vehicleRequirement,
+            status: this.state.status,
+            clientId: this.state.clientId,
+            driverId: this.state.driverId
         };
         axios.post(
-            "/signup/client",
-            newClient
+            "/submit-order",
+            newOrder
         ).then((response) => {
-            // Alert the user their first and last name, clear `this.state.firstName` and `this.state.lastName`, clearing the inputs
-            //alert(`Hello ${this.state.firstName} ${this.state.lastName}`);
-            this.setState({
-                firstName: "",
-                lastName: "",
-                email: "",
-                phoneNumber: "",
-                username: "",
-                password:""
+            console.log(JSON.stringify(response));
+            thisState.setState({
+                startLocation: "",
+                endLocation: "",
+                loadDescritpion: "",
+                vehicleRequirement: "",
+                status: "",
+                clientId: "",
+                driverId: ""
             });
-            console.log(response);
         }).catch((error) => {
-            console.log(error);
-        });;
+            console.log(JSON.stringify(error));
+        });
     };
     
-      render() {
-          // Notice how each input has a `value`, `name`, and `onChange` prop
-          return (
-              <div>
-                  <h1>Signup for Pickup</h1>
-                  <h3>{(this.state.firstName || this.state.lastName)? `Hello ${this.state.firstName} ${this.state.lastName}` : null}</h3>
-                  <form className="form">
-                      <p>Personal Information</p>
-                      <input
-                        value={this.state.firstName}
-                        name="firstName"
-                        onChange={this.handleInputChange}
-                        type="text"
-                        placeholder="First Name"
-                        autoComplete="given-name"
-                        required
-                      />
-                      <input
-                        value={this.state.lastName}
-                        name="lastName"
-                        onChange={this.handleInputChange}
-                        type="text"
-                        placeholder="Last Name"
-                        autoComplete="family-name"
-                        required
-                      />
-                      <input 
-                        value={this.state.email}
-                        name="email"
-                        onChange={this.handleInputChange}
-                        type="email"
-                        placeholder="Email"
-                        autoComplete="email"
-                        required
-                      />
-                      <input 
-                        value={this.state.phoneNumber}
-                        name="phoneNumber"
-                        onChange={this.handleInputChange}
-                        type="tel"
-                        placeholder="Phone: XXX-XXX-XXXX"
-                        autoComplete="tel"
-                        required
-                      />
-                      <p>Account Information</p>
-                      <input
-                        value={this.state.username}
-                        name="username"
-                        onChange={this.handleInputChange}
-                        type="text"
-                        placeholder="Username"
-                        autoComplete="username"
-                        required
-                      />
-                      <input
-                        value={this.state.password}
-                        name="password"
-                        onChange={this.handleInputChange}
-                        type="password"
-                        placeholder="Password"
-                        autoComplete="new-password"
-                        required
-                      />
-                      <button onClick={this.handleFormSubmit}>Submit</button>
-                  </form>
-              </div>
-          );
-      }
+    render() {
+        // Notice how each input has a `value`, `name`, and `onChange` prop
+        return (
+            <div>
+                {/* <p>
+                    Hello {this.state.firstName} {this.state.lastName}
+                </p>*/}
+                <form className="form">
+                    <Row>
+                        <Input
+                          value={this.state.startLocation}
+                          name="startLocation"
+                          onChange={this.handleInputChange}
+                          type="text"
+                          placeholder="Pick-Up Location"
+                        />
+                        <Input
+                          value={this.state.endLocation}
+                          name="endLocation"
+                          onChange={this.handleInputChange}
+                          type="text"
+                          placeholder="Drop-off Location"
+                        />
+                    </Row>
+                    <Input 
+                      value={this.state.vehicleRequirement}
+                      name="vehicleRequirement"
+                      onChange={this.handleInputChange}
+                      type="text"
+                      placeholder="Vehicle Requirement"
+                    />
+                    <Input 
+                      value={this.state.vehicleRequirement}
+                      name="loadDescription"
+                      onChange={this.handleInputChange}
+                      type="text"
+                      placeholder="Load Description"
+                    />
+                    <Button onClick={this.handleFormSubmit}>Submit</Button>
+                </form>
+            </div>
+        );
+    }
 }
 
-export default DriverSignup;
+
+export default ClientOrder;
