--- client/src/pages/DispatchOverview/index.js.orig
@@ -1,13 +1,0 @@
-import React from "react";
-
-function DispatchOverview() {
-    return(
-        <div>
-            <h1>Dispatch Main Overview</h1>
-            <p>Map goes here.</p>
-            <p>Order go here.</p>
-        </div>
-    );
-}
-
-export default DispatchOverview;
