--- client/src/pages/DriverOrder/DriverOrder.jsx.orig
@@ -1,12 +1,0 @@
-import React from "react";
-
-function DriverOrder() {
-    return(
-        <div>
-            <h1>Driver View of Order No. 12343</h1>
-            <p>Order info goes here.</p>
-        </div>
-    );
-}
-
-export default DriverOrder;
