--- client/src/pages/DriverOrder/index.js~af3d2868ffa39b070570cbfcdb734ddbb4e17841
@@ -1,12 +1,0 @@
-import React from "react";
-
-function DriverOrder() {
-    return(
-        <div>
-            <h1>Driver View of Order No. 12343</h1>
-            <p>Order info goes here.</p>
-        </div>
-    );
-}
-
-export default DriverOrder;
