--- client/src/pages/DriverOverview/index.js.orig
@@ -1,13 +1,0 @@
-import React from "react";
-
-function DriverOverview() {
-    return(
-        <div>
-            <h1>Driver Main Overview</h1>
-            <p>Map goes here.</p>
-            <p>Customer order list goes here.</p>
-        </div>
-    );
-}
-
-export default DriverOverview;
