--- client/src/pages/PickupHome/index.js.orig
@@ -1,13 +1,0 @@
-import React from "react";
-
-function PickupHome() {
-    return(
-        <div>
-            <h1>Pickup!</h1>
-            <p>Hotshot management system.</p>
-            <p>sign up</p>
-        </div>
-    );
-}
-
-export default PickupHome;
