--- package_BACKUP_36832.json
@@ -1,40 +1,0 @@
-{
-  "name": "pickup-app",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "start": "if-env NODE_ENV=production && yarn run start:prod || yarn run start:dev",
-    "start:prod": "node server.js",
-    "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"yarn run client\"",
-    "client": "cd client && yarn run start",
-    "install": "cd client && yarn install",
-    "build": "cd client && yarn run build",
-    "heroku-postbuild": "yarn run build"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@material-ui/core": "^3.9.1",
-    "bcrypt-nodejs": "^0.0.3",
-    "bootstrap": "^4.2.1",
-    "express": "^4.16.4",
-    "express-session": "^1.15.6",
-    "if-env": "^1.0.4",
-    "mysql2": "^1.6.4",
-    "passport": "^0.4.0",
-    "passport-local": "^1.0.0",
-    "react": "^16.7.0",
-    "react-bootstrap": "^1.0.0-beta.5",
-    "react-dom": "^16.7.0",
-    "react-mdl": "^1.11.0",
-    "react-scroll": "^1.7.10",
-    "react-stickynode": "^2.1.1",
-    "sequelize": "^4.42.0",
-    "socket.io": "^2.2.0"
-  },
-  "devDependencies": {
-    "concurrently": "^4.1.0"
-  }
-}
