+++ routes/api-routes.js
@@ -68,6 +68,26 @@ module.exports = function(app) {
   //     });
   //   }
   // });
+    //order submit route
+    app.post("/submit-order", (req, res) => {
+        const newOrder = req.body;
+        console.log(newOrder);
+        // db.Order.create({
+        //     start_location: newOrder.startLocation,
+        //     end_location: newOrder.endLocation,
+        //     load_description: newOrder.loadDescritpion,
+        //     vehicle_requirement: newOrder.vehicleRequirement,
+        //     status: newOrder.status,
+        //     client_id: newOrder.clientId,
+        //     driver_id: newOrder.driverId
+        // }).then((response) => {
+        //     console.log(response);
+            // res.json({success: "New order was successfully submitted."});
+        // }).catch((error) => {
+        //     console.log(error);
+            // res.json({error: "There was an error creating submitting the new order."});
+        // });
+    });
   
     // client signup route
     app.post("/client-signup", (req, res) => {
